requirements.txt
langchain
openai
python-dotenv
pydantic[dotenv]